// src/savvy/enums/RecordType.ts
/**
 * @description Record Type Enum
 * @author Sheng Huang
 * @date 2021/11/29
 */

export enum RecordTypeEnum {
    MERCHANDISING = 'Merchandising',
    PRESELL = 'Pre-Sell',
    DELIVERY = 'Delivery',
    SALES = 'Sales',
    TASK = 'Task',
    ORDER = 'Order',
    ACTIVE_PRODUCT = 'Active_Product'
}
export enum VisitListRecordTypeEnum {
    VisitListGroup = 'Visit_List_Group',
    ScheduleVisitGroup = 'Schedule_Visit_Group',
    SalesDailyVisitList = 'Sales_Daily_Visit_List',
    MerchDailyVisitList = 'Daily_Visit_List',
    DeliveryDailyVisitList = 'Delivery_Daily_Visit_List'
}