// src/savvy/enums/VisitList.ts
/**
 * @description Enums of visit list
 * @author Hao Xiao
 * @date 2021-08-02
 */

export enum VisitListStatus {
    NOT_STARTED = 'Not Started',
    ACCEPTED = 'Accepted',
    IN_PROGRESS = 'In Progress',
    COMPLETED = 'Completed',
    CANCELLED = 'Cancelled',
    CANCELING = 'Canceling',
    FAILED = 'Failed',
    PRE_PROCESSED = 'Pre-Processed'
}